+++ packages/ckeditor5-html-support/src/augmentation.ts
@@ -7,7 +7,6 @@ import type {
 	GeneralHtmlSupport,
 	DataFilter,
 	DataSchema,
-	FullPageConfig,
 	GeneralHtmlSupportConfig,
 	CodeBlockElementSupport,
 	CustomElementSupport,
@@ -34,14 +33,6 @@ declare module '@ckeditor/ckeditor5-core' {
 		 * Read more in {@link module:html-support/generalhtmlsupportconfig~GeneralHtmlSupportConfig}.
 		 */
 		htmlSupport?: GeneralHtmlSupportConfig;
-
-		/**
-		 * The configuration of the Full page editing feature.
-		 * Introduced by the {@link module:html-support/fullpage~FullPage} feature.
-		 *
-		 * Read more in {@link module:html-support/fullpageconfig~FullPageConfig}.
-		 */
-		fullPage?: FullPageConfig;
 	}
 
 	interface PluginsMap {
