+++ packages/ckeditor5-html-support/src/generalhtmlsupportconfig.ts
@@ -93,4 +93,97 @@ export interface GeneralHtmlSupportConfig {
 	 * @default false
 	 */
 	preserveEmptyBlocksInEditingView?: boolean;
+
+	/**
+	 * The configuration of the Full page editing feature.
+	 * The option is used by the {@link module:html-support/fullpage~FullPage} feature.
+	 *
+	 * ```ts
+	 * ClassicEditor
+	 * 	.create( {
+	 * 		htmlSupport: {
+	 * 			fullPage: ... // Full page feature config.
+	 * 		}
+	 * 	} )
+	 * 	.then( ... )
+	 * 	.catch( ... );
+	 * ```
+	 */
+	fullPage?: FullPageConfig;
+}
+
+/**
+ * The configuration of the Full page editing feature.
+ */
+export interface FullPageConfig {
+
+	/**
+	 * Whether the feature should allow the editor to render styles from the `<head>` section of editor data content.
+	 *
+	 * When set to `true`, the editor will render styles from the `<head>` section of editor data content.
+	 *
+	 * ```ts
+	 * ClassicEditor
+	 * 	.create( {
+	 * 		htmlSupport: {
+	 * 			fullPage: {
+	 * 				allowRenderStylesFromHead: true
+	 * 			}
+	 * 		}
+	 * 	} )
+	 * 	.then( ... )
+	 * 	.catch( ... );
+	 * ```
+	 *
+	 * @default false
+	 */
+	allowRenderStylesFromHead?: boolean;
+
+	/**
+	 * Callback used to sanitize the CSS provided by the user in editor content
+	 * when option `htmlSupport.fullPage.allowRenderStylesFromHead` is set to `true`.
+	 *
+	 * We strongly recommend overwriting the default function to avoid XSS vulnerabilities.
+	 *
+	 * The function receives the CSS (as a string), and should return an object
+	 * that matches the {@link module:html-support/generalhtmlsupportconfig~CssSanitizeOutput} interface.
+	 *
+	 * ```ts
+	 * ClassicEditor
+	 * 	.create( editorElement, {
+	 * 		htmlSupport: {
+	 * 			fullPage: {
+	 * 				allowRenderStylesFromHead: true,
+	 *
+	 * 				sanitizeCss( CssString ) {
+	 * 					const sanitizedCss = sanitize( CssString );
+	 *
+	 * 					return {
+	 * 						css: sanitizedCss,
+	 * 						// true or false depending on whether the sanitizer stripped anything.
+	 * 						hasChanged: ...
+	 * 					};
+	 * 				}
+	 * 			}
+	 * 		}
+	 * 	} )
+	 * 	.then( ... )
+	 * 	.catch( ... );
+	 * ```
+	 *
+	 */
+	sanitizeCss?: ( css: string ) => CssSanitizeOutput;
+}
+
+export interface CssSanitizeOutput {
+
+	/**
+	 * An output (safe) CSS that will be inserted into the document.
+	 */
+	css: string;
+
+	/**
+	 * A flag that indicates whether the output CSS is different than the input value.
+	 */
+	hasChanged: boolean;
 }
