+++ packages/ckeditor5-html-support/src/htmlpagedataprocessor.ts
@@ -15,6 +15,11 @@ import { HtmlDataProcessor, UpcastWriter, type ViewDocumentFragment } from 'cked
  */
 export default class HtmlPageDataProcessor extends HtmlDataProcessor {
 	/**
+	 * Contains the `documentElement` property of the `Document` interface.
+	 */
+	public parsedDocument: HTMLElement | null = null;
+
+	/**
 	 * @inheritDoc
 	 */
 	public override toView( data: string ): ViewDocumentFragment {
@@ -52,6 +57,8 @@ export default class HtmlPageDataProcessor extends HtmlDataProcessor {
 
 		// Using the DOM document with body content extracted as a skeleton of the page.
 		writer.setCustomProperty( '$fullPageDocument', domFragment.ownerDocument.documentElement.outerHTML, viewFragment );
+
+		this.parsedDocument = domFragment.ownerDocument.documentElement;
 
 		if ( docType ) {
 			writer.setCustomProperty( '$fullPageDocType', docType, viewFragment );
