+++ packages/ckeditor5-html-support/src/index.ts
@@ -14,7 +14,6 @@ export { default as HtmlComment } from './htmlcomment.js';
 export { default as FullPage } from './fullpage.js';
 export { default as HtmlPageDataProcessor } from './htmlpagedataprocessor.js';
 export { default as EmptyBlock } from './emptyblock.js';
-export type { FullPageConfig } from './fullpageconfig.js';
 export type { GeneralHtmlSupportConfig } from './generalhtmlsupportconfig.js';
 export type { default as CodeBlockElementSupport } from './integrations/codeblock.js';
 export type { default as CustomElementSupport } from './integrations/customelement.js';
